Adding travis and appveyor cotinuous integration tests
This is an automated pull request that enables continous integration testing for openFrameworks
addons. We've done our best to detect that this repository actually contains an openFrameworks
addon but if it doesn't or this addon is not maintained anymore please just close or ignore this PR
and no further communications will be sent to this repository.

Before merging this PR you'll need to create an account on https://travis.org and
https://ci.appveyor.com using your github account and enable the tests for this addon there.

Once you have created an account at travis.org you can enable tests for this addon at
https://travis.org/arturoc/ofxTween

And for windows at https://ci.appveyor.com/projects/new

This PR includes test files for travis and appveyor ci services, once you merge it any new commit
to your repository will compile this addon for every supported platform against the latest
openFrameworks.

This new files can be customized but we recommend to not touch them much except for commenting or
uncommenting the different platfoms to test. That way we can ensure that the addon will work for
any standard setup. The testing can be further customized (for example to install some dependencies
in certain platforms by using some scripts in scripts/platforms there's more instructions on how
to use them in the .travis.yml and appveyor.yml files in this PR.

If you have any other doubt you can answer directly in this PR or ask in the openFrameworks forum